requirements.txt
gitdb==4.0.10
GitPython==3.1.30
lizard==1.17.10
numpy==1.25.2
pandas==2.1.0
PyDriller==2.4.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
six==1.16.0
smmap==5.0.0
types-pytz==2023.3.0.1
tzdata==2023.3
